backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
# psycopg2-binary==2.9.9  # Commented out for now - will use SQLite for MVP
alembic==1.12.1

# Data fetching and processing
yfinance==0.2.28
feedparser==6.0.10
pandas==2.1.4
numpy==1.25.2

# Machine Learning
scikit-learn==1.3.2

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Background tasks
apscheduler==3.10.4

# CORS for frontend communication
python-multipart==0.0.6 